refactor: 테이블 뷰 데이터 소스 분리
